REF move maybe_get_random_name to under civis.io
