+++ src/civis/io/_databases.py
@@ -1,7 +1,7 @@
 import logging
 
 from civis import APIClient
-from civis._utils import maybe_get_random_name
+from civis.io._utils import maybe_get_random_name
 from civis.futures import CivisFuture
 
 log = logging.getLogger(__name__)
