+++ src/civis/io/_utils.py
@@ -1,0 +1,7 @@
+import uuid
+
+
+def maybe_get_random_name(name):
+    if not name:
+        name = uuid.uuid4().hex
+    return name
